+++ pyscriptjs/tests/integration/test_01_basic.py
@@ -219,16 +219,15 @@ class TestBasic(PyScriptTest):
         assert self.console.log.lines[-1] == "hello from foo"
 
     def test_py_script_src_not_found(self):
-        self.pyscript_run(
-            """
-            <py-script src="foo.py"></py-script>
-            """
-        )
+        with pytest.raises(JsErrors) as exc:
+            self.pyscript_run(
+                """
+                <py-script src="foo.py"></py-script>
+                """
+            )
         assert self.PY_COMPLETE in self.console.log.lines
 
         assert "Failed to load resource" in self.console.error.lines[0]
-        with pytest.raises(JsErrors) as exc:
-            self.check_js_errors()
 
         error_msgs = str(exc.value)
 
@@ -372,3 +371,17 @@ class TestBasic(PyScriptTest):
         btn.click()
         assert self.console.log.lines[-1] == "hello world!"
         assert self.console.error.lines == []
+
+    def test_py_mount_shows_deprecation_warning(self):
+        # last non-deprecated version: 2023.03.1
+        self.pyscript_run(
+            """
+            <div id="foo" py-mount></div>
+            """
+        )
+        banner = self.page.locator(".alert-banner")
+        expected_message = (
+            'The "py-mount" attribute is deprecated. '
+            + "Please add references to HTML Elements manually in your script."
+        )
+        assert banner.inner_text() == expected_message